+++ src/components/app/App.jsx
@@ -1,6 +1,12 @@
 import React from 'react';
 import Profile from '../profiles/Profile';
+import ProfileSearch from '../profiles/ProfileSearch';
 
 export default function App() {
-  return <Profile />;
+  return (
+    <>
+      <ProfileSearch />
+      <Profile />
+    </>
+  );
 }
