+++ src/components/profiles/Profile.jsx
@@ -1,48 +1,36 @@
 /* eslint-disable keyword-spacing */
 /* eslint-disable max-len */
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { fetchUser, setLoading, setSearch } from '../../actions/githubActions';
-import ProfileSearch from './ProfileSearch';
+import { useSelector } from 'react-redux';
 
 const Profile = () => {
-  const user = (state => state.user);
+  const user = useSelector(state => state.user);
   const loading = useSelector(state => state.loading);
-  const dispatch = useDispatch();
 
-  const handleSubmit = userName => {
-    dispatch(setLoading(true));
-    dispatch(fetchUser(userName));
-  };
+  if (loading) {
+    return <h1>Loading...</h1>;
 
-  const handleChange = ({ target }) => {
-    dispatch(setSearch(target.value));
-  };
+  } else if (user === {}) {
+    const repoElements = user.userRepos.map(repo => (
+      <li key={repo.name}>
+        <p>Name: {repo.name}</p>
+        <a href={repo.url}>Link: {repo.url}</a>
+      </li>
+    ));
 
-  const repoElements = user.userRepos.map(repo => (
-    <li key={repo.name}>
-      <p>Name: {repo.name}</p>
-      <a href={repo.url}>Link: {repo.url}</a>
-    </li>
-  ));
-
-  if (loading) return <h1>Loading...</h1>;
-
-  if (user === {}) return <ProfileSearch onChange={handleChange} onSubmit={handleSubmit} />;
-
-  return (
-    <>
-      <ProfileSearch onChange={handleChange} onSubmit={handleSubmit} />
-      <h1>{user.userProfile.name}</h1>
-      <a href={user.userProfile.url}>Link: {user.userProfile.url}</a>
-      <p>Follower Count: {user.userProfile.followers}</p>
-      <p>Following {user.userProfile.followers} Users</p>
-      <ul>
-        Repositories:
-        {repoElements}
-      </ul>
-    </>
-  );
+    return (
+      <>
+        <h1>{user.userProfile.name}</h1>
+        <a href={user.userProfile.url}>Link: {user.userProfile.url}</a>
+        <p>Follower Count: {user.userProfile.followers}</p>
+        <p>Following {user.userProfile.followers} Users</p>
+        <ul>
+          Repositories:
+          {repoElements}
+        </ul>
+      </>
+    );
+  } else { return <p>Please Make a Search</p>; }
 };
 
 export default Profile;
