+++ src/components/profiles/ProfileSearch.jsx
@@ -1,10 +1,24 @@
 import React from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { fetchUser, setLoading, setSearch } from '../../actions/githubActions';
 
-const ProfileSearch = (handleChange, handleSubmit) => {
+const ProfileSearch = () => {
+  const dispatch = useDispatch();
+  const userName = useSelector(state => state.search);
+
+  const handleSubmit = () => {
+    dispatch(setLoading(true));
+    dispatch(fetchUser(userName));
+  };
+  
+  const handleChange = ({ target }) => {
+    dispatch(setSearch(target.value));
+  };
+
   return (
     <>
       <input onChange={handleChange} />
-      <button onClick={handleSubmit} />
+      <button onClick={handleSubmit}>Search Github Profiles</button>
     </>
   );
 };
